+++ _goldens/test/_files/directives/directive_wrapper.dart
@@ -86,36 +86,3 @@ class TestFooComponent {
 class MyInjectableClass {
   String get title => 'hello';
 }
-
-@Directive(
-  selector: '[fastDirective]',
-)
-class FastDirective extends ComponentState {
-  Element element;
-
-  @HostBinding('attr.data-msg')
-  String msg;
-  String _prevValue;
-
-  FastDirective(this.element);
-
-  @Input()
-  set name(String value) {
-    if (_prevValue == value) return;
-    _prevValue = value;
-    setState(() => msg = 'Hello $value');
-  }
-}
-
-@Component(
-  selector: 'directive-container',
-  template: r'''
-    <div class="target1" fastDirective [name]="finalName"></div>
-    <div class="target2" fastDirective [name]="nonFinal"></div>
-  ''',
-  directives: [FastDirective],
-)
-class DirectiveContainerTest {
-  final String finalName = "xyz";
-  String nonFinal = "abc";
-}