+++ angular/lib/src/compiler/template_parser/ast_template_parser.dart
@@ -906,9 +906,9 @@ class _ElementFilter extends ast.RecursiveTemplateAstVisitor<Null> {
       // TODO: Add a flag to upgrade this to an error.
       final warning = astNode.sourceSpan.message(
         ''
-            'Ignoring <${astNode.name}>, as this element is unsafe to bind in '
-            'a template without proper sanitization. This may become an error '
-            'in future versions of AngularDart. See $_securityUrl for details.',
+        'Ignoring <${astNode.name}>, as this element is unsafe to bind in '
+        'a template without proper sanitization. This may become an error '
+        'in future versions of AngularDart. See $_securityUrl for details.',
       );
       logWarning(warning);
       return null;